+++ types/easymde.d.ts
@@ -146,6 +146,7 @@ declare namespace EasyMDE {
         name: string;
         action: string | ((editor: EasyMDE) => void);
         className: string;
+        text?: string;
         title: string;
         noDisable?: boolean;
         noMobile?: boolean;