+++ CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to EasyMDE will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.x] - 2024-07-02
+#### Added
+- Add indent and outdent toolbar buttons
+- Add characters count to status bar
+- Add options to change the statusbar words for autosave, characters, lines and words
+
+#### Fixed
+- Fix upload image
+- Fix unused iconClassMap option
+- Fix ToolbarIcon type
+- Fixed the toggling behaviour in toggleLinks
+
 ## [1.0.x] - Up to 2024-07-01
 #### Started
 - Create a fork from EasyMDE
 
-#### Fuxed
+#### Fixed
 - Bugs fixed